Refactor BlockControls to override only slot implementation
