+++ packages/block-editor/src/components/block-controls/slot.native.js
@@ -1,0 +1,33 @@
+/**
+ * WordPress dependencies
+ */
+import { useContext } from '@wordpress/element';
+import {
+	__experimentalToolbarContext as ToolbarContext,
+	ToolbarGroup,
+} from '@wordpress/components';
+
+/**
+ * Internal dependencies
+ */
+import groups from './groups';
+
+export default function BlockControlsSlot( { group = 'default', ...props } ) {
+	const accessibleToolbarState = useContext( ToolbarContext );
+	const Slot = groups[ group ].Slot;
+
+	if ( group === 'default' ) {
+		return <Slot { ...props } fillProps={ accessibleToolbarState } />;
+	}
+
+	return (
+		<Slot { ...props } fillProps={ accessibleToolbarState }>
+			{ ( fills ) => {
+				if ( ! fills.length ) {
+					return null;
+				}
+				<ToolbarGroup>{ fills }</ToolbarGroup>;
+			} }
+		</Slot>
+	);
+}
