+++ packages/components/src/slot-fill/index.js
@@ -44,6 +44,7 @@ export function createSlotFill( name ) {
 
 	const SlotComponent = ( props ) => <Slot name={ name } { ...props } />;
 	SlotComponent.displayName = name + 'Slot';
+	SlotComponent.__unstableName = name;
 
 	return {
 		Fill: FillComponent,
